refactor!: Reorganize files
BREAKING CHANGE:
Moves all source files into a subdirectory named src within the PLCC project.
The path to this src directory is what the LIBPLCC environment variable must
point to, and is also the path that must be added to the PATH environment
variable.  When upgrading, these environment variables must be manually
adjusted to point to the new src directory.