chore(shell): Pin Alpine packages to approximate version numbers
This commit only applies to the shell that is used to manually
test PLCC (in ./shell).

Alpine does not keep old package versions. So if you pin to a
specific package version, your build will break if Alpine
updates to a new version of the package.

To allow package versions to varry, but maintain some level
of consistency in builds, we're pinning to approximate versions
instead. For example, instead of

	openjdk=11.0.7-rc0

We now pin to

	openjdk~=11.0

This allows the rest of the version number to spin as needed.
In theory, allowing patch updates should not break our build.