build(shell): Remove build args to enable caching
ARG lines in Dockerfile invalidates build caches from that point on.
The existing ARG lines were used as constants so that version numbers
could be specified at the top of the file. But their adverse effect on
build performance does not outweigh their benefit in design. So this
commit inlines them.