cd: enable continuous delivery from master
This commit enables continuous delivery. Whenever a commit is pushed
to master (for example, when a pull-request is merged into master),
the same test shell used in continuous integration will again be built
and tested. But this time, that image will also be pushed to
ourplcc/plcc-in-docker:latest on DockerHub. In short, this will keep
the plcc-in-docker:latest Docker image in synch with the contents of
master.

Note that the release process is still manual. To release, one still
needs to update the change log, bump the version number, tag the commit
with a release tag, and tag the corresponding image with the appropriate
release tags in DockerHub. Although some of this is automated by scripts
in release, it is not performed automaticly. That is, a human still has
to initiate the release. This commit makes the current "release
candidate" image available in DockerHub under the "latest" tag.

Last note, after this pull-request is merged, this commit will likely be
followed by several commits to adjust the workflow, as the only way to
test the workflow is by installing it and seeing what happens.