refactor(ci): fix build failure (#98)
Checks were added to prevent folks from breaking their system's
Python installation by updating it or adding libraries to it.

Our Dockerfile was doing this. So now we don't. We still make
sure that Python 3 is installed. But we don't try to upgrade it
or pip.

In the future, we may want to figure out how to use something
like venv to install a particular version of Python that PLCC will
use, that is independent of the system's Python.

Closes #97