refactor: enable Python when grammar has 4 sections
This is in anticipation of a new hybrid Java/Python mode
in which the parser runs in Java and the semantics rep-loop
runs in Python.

In this commit, when an optional 4th section is detected,
plcc.py generates ParseJsonAst.json. Thus allowing the
parser to to generate JSON ASTs when the `--json_ast`
flag is passed. It also generates experimental Python files.

---

Closes #95

Co-authored-by: Reed Everis <reednorrish@gmail.com>
Co-authored-by: Akshar Patel <aksharpatel1233@gmail.com>
Co-authored-by: Stoney Jackson <dr.stoney@gmail.com>