fix: report version number from git tag (#134)
First, because __main__.py was moved, it was looking for .git/ in the wrong location.
That's now fixed.

Second, the test for the --version option allowed for the "Unknown" result.
This is now a failure so that we can detect this regression in the future.

Third, CI was not cloning the complete history. So when the Docker images
where built, and the repository was copied into the image, the version tags
where not in the history. The GitHub actions have been updated to include
the complete Git history when it checks out the repository. Long run,
this could slow CI and/or make the images much larger than they need to be.
Short run, I don't think we are in any danger of that. If we notice this, then
we'll need to set the "depth" in our GitHub actions to something like 100.

Last, Git was not happy with the ownership of the repositories copied into
the Docker images. This commit marks these repositories as safe by running
"git config --global --add safe.directory /plcc" when building containers.